Changed a lot
Using TWEEN.js I managed to make the camera movement
smoother when pressing the gui buttons, which were also
moved into folders to make a distinction between options.
Added and used a custom class called rubiksCube to represent
the cube using OOP, using the fetch api managed to make it
that with a press of a button the cube gets colored properly.